[14.0][IMP] product_pack_type: Improve tests
