+++ product_pack_type/tests/test_product_pack_type.py
@@ -2,23 +2,14 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from odoo.tests.common import SavepointCase
 
-from odoo.addons.product_pack.tests.common import ProductPackCommon
+from .common import ProductPackTypeCommon
 
 
-class TestProductPack(ProductPackCommon, SavepointCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.pack_type_obj = cls.env["product.pack.type"]
-
-    @classmethod
-    def _create_type(cls, vals=None):
-        if not vals:
-            vals = {}
-        vals.update({"name": "CPUs"})
-        return cls.pack_type_obj.create(vals)
-
+class TestProductPack(ProductPackTypeCommon, SavepointCase):
     def test_product_pack_type_count(self):
+        # Create a pack type
+        # Affect a pack product to that type
+        # Check count
         pack_type = self._create_type()
         self.assertEqual(0, pack_type.product_pack_count)
         self.cpu_detailed.pack_type_id = pack_type
